Move routes/pre-handlers for consistency with other modules.  Add newline to check-answers.njk
