+++ src/internal/modules/returns-notifications/routes.js
@@ -1,0 +1,6 @@
+'use strict';
+
+module.exports = {
+  ...require('./routes/batch-notifications'),
+  ...require('./routes/paper-forms')
+};
