--- src/internal/modules/returns-notifications/routes/index.js
@@ -1,6 +1,0 @@
-'use strict';
-
-module.exports = {
-  ...require('./batch-notifications'),
-  ...require('./paper-forms')
-};
